# swagger/apis/swagger.yaml
...
swagger: '2.0'
info:
  description: 'printer public APIs documentation'
  version: 1.0.6
  title: ePRINTit SAAS
host: api.eprintitsaas.org/public/
schemes:
  - https
security:
  - APIKeyHeader: []
paths:
...
  "/requestToken":
    get:
      tags:
        - kiosk
      summary: get the token over the IoT to the kiosk app
      description: ''
      operationId: requestToken
      produces:
        - application/json
      parameters:
        - name: thingTagId
          in: query
          description: 'thing tag id of the kiosk app'
          required: true
          type: string
        - in: header
          name: apiKey
          type: string
          required: true
        - in: header
          name: subdomain
          type: string
          required: false
      responses:
        '200':
          description: successful operation
          type: json
...